+++ tests/Mock/MockTableViewWithRedirectActions.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace LaravelViews\Test\Mock;
+
+use LaravelViews\Actions\Action;
+use LaravelViews\Actions\RedirectAction;
+use LaravelViews\Filters\Filter;
+
+class MockTableViewWithRedirectActions extends MockTableView
+{
+    protected function actionsByRow()
+    {
+        return [
+            new RedirectAction('user', 'See user', 'eye'),
+            new RedirectAction('user.edit', 'Edit user', 'pencil'),
+        ];
+    }
+}